Fixed Calibration Issue
In initial piloting of the task, I discovered that the pure tone played during the task would be approximately 3-4dB quieter than the tone played during the calibration. I thought this was due to a difference in the way the tone was being played across tasks (initially the calibration tone was generated within psychopy, while the test tone was generated using an external application). The fix was to increase the dB in the calibration files to 117 rather than 113. After some thinking, I realised that the under-reading of the tone volume in the testing session compared to the calibration session was due to tone duration. The tones in the calibration session are played indefintiely, allowing the dB SPL meter to get a more accurate read of the true dB SPL. I have fixed this error by making the calibration value 113 across all the files.